+++ src/components/organisms/Navbar/Navbar/Navbar.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import styled, { css } from 'styled-components';
 import Headroom from 'react-headroom';
 import { faBars } from '@fortawesome/free-solid-svg-icons';
@@ -26,7 +26,8 @@ export interface NavigationItem {
   href?: string;
   'data-automation'?: string;
   onClick?: (event?: React.MouseEvent<HTMLAnchorElement>) => void;
-  children?: NavigationItem[];
+  isDropdownHeading?: boolean;
+  children?: Exclude<NavigationItem, 'children'>[];
 }
 export interface NavbarLinksListProps {
   /**
@@ -78,6 +79,7 @@ const PageNavigation = styled.header<PageNavigationProps>`
   top: 0;
   left: 0;
   right: 0;
+  z-index: 1;
 
   ${minMedia.desktop`
     ${css`
@@ -219,9 +221,9 @@ const HamburgerMenu = styled.aside<HamburgerContainerProps>`
 
   display: flex;
   flex-direction: column;
-  min-height: 100vh;
+  min-height: calc(100vh - ${mobileNavbarHeight}px);
   width: 100%;
-  padding: ${spacing[8]} ${spacing[4]} 0;
+  padding: ${spacing[8]} ${spacing[4]};
 
   background: ${colors.white};
 
@@ -229,6 +231,7 @@ const HamburgerMenu = styled.aside<HamburgerContainerProps>`
   transform: ${({ open }) => (open ? 'translateX(0)' : 'translateX(-100%)')};
   z-index: 2;
   overflow-y: auto;
+  -webkit-overflow-scrolling: touch;
 `;
 
 const LargeDeviceNavbar = styled.div`
@@ -251,12 +254,8 @@ const SmallDeviceNavbar = styled.div`
   `}
 `;
 
-export const NavbarLinksListLink = ({
-  item: { label, href, onClick, ...rest },
-  index,
-  props,
-}: NavbarLinksListLinkProps) => (
-  <NavbarLink key={`navbar-link-${index}`} href={href} onClick={onClick} {...props} {...rest}>
+export const NavbarLinksListLink = ({ item: { label, onClick, ...rest }, index, props }: NavbarLinksListLinkProps) => (
+  <NavbarLink key={`navbar-link-${index}`} {...props} {...rest}>
     {label}
   </NavbarLink>
 );
@@ -274,6 +273,14 @@ const NavbarWrapper: React.FC<NavbarProps> = ({
   const { width } = useViewport();
   const overThreshold = useScrollThreshold(20);
   const [open, setOpen] = useState<boolean>(false);
+
+  useEffect(() => {
+    if (open) {
+      document.body.classList.add('nav-open');
+    } else {
+      document.body.classList.remove('nav-open');
+    }
+  }, [open]);
 
   return (
     <>